refactor: Rename verbs/ to verb/
